[#91768006] Selects more efficiently the files to pack
At the beginning, we retrieved all the files that were in the directory of
the part.cfg (part description file) of the parts that were used in the
current craft. To save space (especially for the cases where a directory
contains several parts), we check deeper in the files to retrieve
only the files that are actually used (.cfg + .mu + .mbm(s)).