[#91768005] Adds MBM to PNG conversion
MBM is the format used by KSP for models'textures. MBM is a derivative of
PBM (Portable Bit Map) which is heavy and makes the craft archives to be
huge. Texture conversion (MBM -> PNG) is now done by the exporter to
reduce final archive size, which allows the user to be under the size
limit if not PRO.